==> code/notebooks/kevin/01_scrapePDFdata.ipynb <==
Author: Kevin ALBERT

Created: March 2021

Here we need to transform the pdf to csv

!ls -al ../../../data/bronze/Ultratop/Flanders |grep pdf

import pandas as pd
import tabula
import os

files = !ls ../../../data/bronze/Ultratop/Flanders |grep pdf
files

# files = ['Ultratop - Singles - 02_01_2021.pdf']

# files = os.listdir('../../../data/bronze/Ultratop/Flanders/')
# files

ultraTopPDF = pd.DataFrame()
for file in files:
    data = tabula.read_pdf(f'../../../data/bronze/Ultratop/Flanders/{file}', pages="1", lattice=True)[0]
    data.columns = data.iloc[1]
    data = data.drop([0,1])
    data['fileDate'] = f'{file[21:31]}'
    ultraTopPDF = ultraTopPDF.append(data)

ultraTopPDF.info()

ultraTopPDF.sample(5)

ultraTopPDF.to_parquet("../../../data/bronze/Ultratop/Flanders/ultratopPDFflanders.parquet")

ultraTopPDF.to_csv("../../../data/bronze/Ultratop/Flanders/ultratopPDFflanders.csv", index=False)

### now, for Walonie

files = !ls ../../../data/bronze/Ultratop/Walonie |grep pdf
files

ultraTopPDF = pd.DataFrame()
for file in files:
    data = tabula.read_pdf(f'../../../data/bronze/Ultratop/Walonie/{file}', pages="1", lattice=True)[0]
    data.columns = data.iloc[1]
    data = data.drop([0,1])
    data['fileDate'] = f'{file[21:31]}'
    ultraTopPDF = ultraTopPDF.append(data)

ultraTopPDF.info()

ultraTopPDF.sample(5)

ultraTopPDF.to_parquet("../../../data/bronze/Ultratop/Walonie/ultratopPDFwalonie.parquet")

ultraTopPDF.to_csv("../../../data/bronze/Ultratop/Walonie/ultratopPDFwalonie.csv", index=False)

ultraTopPDF.head(10)

# dfs = tabula.read_pdf("../../../data/bronze/Ultratop/Flanders/Ultratop - Singles - 27_02_2021.pdf",
#                       pages="1",
#                       lattice=True)

# read_pdf returns list of DataFrames
# print(len(dfs))
# df = dfs[0]
# df

# # set a row to acolumn:
# df.columns = df.iloc[1]
# df = df.drop([0,1])

# df

# store local as *.csv files
df.to_parquet("../../../data/bronze/Ultratop/Flanders/Ultratop - Singles - 27_02_2021.parquet")

for
