==> code/notebooks/kevin/03_dataMerging_UltratopXLS-Spotify.ipynb <==
Author: Kevin ALBERT

Created: March 2021

import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript

spotify = pd.read_csv("../../../data/silver/spotify.csv")
ultra   = pd.read_csv("../../../data/silver/ultra.csv")

spotify.columns

ultra.columns

ultra

import pandas_profiling as pp
from IPython.display import Javascript

reportFile = "../../../data/report/spotify_report.html"
# quick report on 100% records (no correlation matrix stuff)
pp.ProfileReport(df=spotify.sample(frac=1),
                 minimal=True,
                 progress_bar=False,
                 correlations={"cramers": {"calculate": False}}).to_file(reportFile)

reportFile = "../../../data/report/ultra_report.html"
# quick report on 100% records (no correlation matrix stuff)
pp.ProfileReport(df=ultra.sample(frac=1),
                 minimal=True,
                 progress_bar=False,
                 correlations={"cramers": {"calculate": False}}).to_file(reportFile)

df = ultra.merge(spotify, how='inner', on='Titre')
df

# df.dropna(subset=["length"])

df.to_parquet("../../../data/platinum/dataset.parquet")
