==> code/notebooks/kevin/04_dataMerging_UltratopPDF-Spotify.ipynb <==
Author: Kevin ALBERT

Created: March 2021

import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript

spotify  = pd.read_csv("../../../data/silver/spotify.csv")
ultraPDFflanders = pd.read_parquet("../../../data/bronze/Ultratop/Flanders/ultratopPDFflanders.parquet")
# ultraPDFwalonie = pd.read_parquet("../../../data/bronze/Ultratop/Walonie/ultratopPDFwalonie.parquet")

spotify.columns

ultraPDFflanders.columns

### investigate

filt = ultraPDFflanders['fileDate'] == '02_01_2021'

ultraPDFflanders.VW 

Date        New   RE   Ranking of New   Ranking of RE
13/3/2021   2     2       45,49
6/3/2021    2     1       46,47            45
3/3


shazam.bel    ultratop




200.             34        34.    13/2021

200             41.        35     6/3/2021

200                               27/2/2021

300             50         50
-top 50
- not top 50


result = prep[["winery","country"]].groupby(['country'])['winery'].nunique()
result = result.rename_axis(['Country']).rename('totalNrWineries').sort_values(ascending=False).reset_index()
result.head(10)

### clean dataset

ultraPDFflanders.head(5)

# 5 columns : date, number of new song, number of reentry song, ranking of new songs , ranking of  reentry song
# For both Flanders and Wallonia
