==> code/notebooks/kevin/05_Spotify_Data_Cleaning.ipynb <==
!pip install spotipy

## Enrich the ultratop XLS dataset with Spotify API

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import keys

!pwd

ultra = pd.read_csv("../../../data/silver/ultra.csv", encoding = 'utf-8')

ultra.reset_index(inplace=True, drop=True)

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id="f93a1fbf63074ec3b110cc58b86957dc",
                                                           client_secret="bee7864fda474acdbfc4789c59d53ec1"))

ultra.head(8)

ultra[['Artiste', 'Titre']].agg(''.join, axis=1).iloc[5]

listx = list(ultra[['Artiste', 'Titre']].agg(''.join, axis=1).replace(u'\xa0', u' '))
print(len(listx))
listx

listx = [text.replace(u'\xa0', u' ') for text in listx]
listx = [text.replace(u' x ', u' ') for text in listx]
listx = [text.replace(u'feat. ', u'') for text in listx]
listx = [text.replace(u'Sch - Kofs - Jul - Naps - Soso Maness - Elams - Solda - ', u'') for text in listx]
listx = [text.replace(u'& Mathieu', u'Mathieu') for text in listx]
listx = [text.lower() for text in listx]
listx = [text.replace(u'& zed', u'zed') for text in listx]
listx = [text.replace(u'& camille', u'camille') for text in listx]
listx = [text.replace(u' / the', u'') for text in listx]
listx = [text.replace(u' [fr] ', u' ') for text in listx]
listx = [text.replace(u' & ', u' and ') for text in listx]
listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
listx = [text.replace(u'suarez maria del rio sarà perché ti amo', u"sara' perché ti amo") for text in listx]
listx = [text.replace(u'and wes', u'') for text in listx]
listx = [text.replace(u'vald gotaga', u'gotaga') for text in listx]
listx = [text.replace(u'jul and sch mother fuck', u'mother fuck') for text in listx]
listx = [text.replace(u'laura tesoro and loïc nottet alex germys strangers', u'loic nottet alex germys strangers') for text in listx]
listx = [text.replace(u'amel bent and hatik 1,2,3', u'amel bent hatik 1,2,3') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]
# listx = [text.replace(u'mika and soprano le cœur holiday', u'mika le coeur holiday') for text in listx]

listx

# ultra = ultra.head(5)
# ultra

# ultra[['Artiste', 'Titre']].agg(''.join, axis=1)

df = pd.DataFrame()
for i, artistTitre in enumerate(listx):
#     print(type(artistTitre)
    print("begin", i, artistTitre)
    data = pd.DataFrame()
    data = ultra.iloc[i].copy()
    
    response = sp.search(q=artistTitre, limit=1)
#     if [title['name'] for title in response['tracks']['items']] == []:
#         artistTitre = artistTitre.replace('feat. ', '')
#         response = sp.search(q=artistTitre, limit=1)
#         print("if 1", i, artistTitre)
#     if [title['name'] for title in response['tracks']['items']] == []:
#         print("if 2", i, artistTitre)
#         artistTitre = artistTitre.replace(' x', ' &')
# #         artistTitre = artistTitre.replace("Amel Bent x Imen Es Jusqu'au bout", "Amel Bent Imen Es Jusqu'au bout")
#         print("if 2 veranderd", i, artistTitre)
#         response = sp.search(q=artistTitre, limit=1)
#         print(i, artistTitre, response)
#     else:
#     print("ga door", i, artistTitre)
    data["CleanedTitle"] = [title['name'] for title in response['tracks']['items']][0]
    data["id"]           = [title['id'] for title in response['tracks']['items']][0]
    data["popularity"]   = [title['popularity'] for title in response['tracks']['items']][0]

    results = sp.audio_features(tracks=[[title['id'] for title in response['tracks']['items']][0]])
    data["danceability"] = results[0]['danceability']
    data["energy"]       = results[0]['energy']
    data["key"]          = results[0]['key']
    data["loudness"]     = results[0]['loudness']
    data["mode"]         = results[0]['mode']
    data["speechiness"]  = results[0]['speechiness']
    data["acousticness"] = results[0]['acousticness']
    data["instrumentalness"] = results[0]['instrumentalness']
    data["liveness"]         = results[0]['liveness']
    data["valence"]          = results[0]['valence']
    data["tempo"]            = results[0]['tempo']
    data["duration_ms"]      = results[0]['duration_ms']
    df = pd.concat([df,data], axis=1)

df = df.T
df

# AMAI t'is gelukt ! pfff

# opslaan die handel
df.to_csv("../../../data/silver/ultraPlusSpotify.csv", index=False)
df.to_parquet("../../../data/silver/ultraPlusSpotify.parquet")

## Enrich the ultratop PDF dataset with Spotify API

ultraPDFflanders = pd.read_parquet("../../../data/bronze/Ultratop/Flanders/ultratopPDFflanders.parquet")

ultraPDFflanders.head(5)

ultraPDFflanders.shape

ultraPDFflanders = ultraPDFflanders.drop_duplicates(subset=['Titel'], keep='last')

ultraPDFflanders.shape

## Merge

# load again the dataset from above
df = pd.read_parquet("../../../data/silver/ultraPlusSpotify.parquet")

df.head(5)

df.shape

df = df.merge(ultraPDFflanders, how="left", left_on="Titre", right_on="Titel").reset_index()
df

## Ok this is quite ok, some Nan valus indicate probably new not found in PDF... does not matter

### save dataset for ML model training

# opslaan die handel
df.to_csv("../../../data/silver/ultraPlusSpotifyplusPDF.csv", index=False)
df.to_parquet("../../../data/silver/ultraPlusSpotifyplusPDF.parquet")
