==> code/notebooks/kevin/06_MachineLearning.ipynb <==
Author: Kevin ALBERT  

Created: March 2021

### Import open-source packages

# environment packages
import platform
import psutil
import os

# other packages
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_colwidth', 100) # default 50, the maximum width in characters of a column
pd.set_option('display.max_columns', 40)   # default 20, the maximum amount of columns in view 
pd.set_option('display.max_rows', 60)      # default 60, the maximum amount of rows in view
import logging
import json
import requests
import joblib

### Import azure machine learning SDK packages

from azureml.core import Workspace, Dataset, Datastore, Run
from azureml.core.experiment import Experiment
from azureml.data.datapath import DataPath
from azureml.core.compute import ComputeTarget, AmlCompute, AksCompute
from azureml.core.model import Model, InferenceConfig
from azureml.train.automl import AutoMLConfig
from azureml.train.automl.run import AutoMLRun
from azureml.widgets import RunDetails
from azureml.core.webservice import Webservice, AciWebservice, AksWebservice
from azureml.exceptions import WebserviceException
from azureml.core.environment import Environment
from azureml.train.estimator import Estimator
from azureml.core.conda_dependencies import CondaDependencies
from azureml.train.hyperdrive.runconfig import HyperDriveConfig
from azureml.train.hyperdrive.sampling import RandomParameterSampling, GridParameterSampling
from azureml.train.hyperdrive.run import PrimaryMetricGoal
from azureml.train.hyperdrive.parameter_expressions import choice
from azureml.core.runconfig import RunConfiguration
from azureml.pipeline.core import PipelineData, Pipeline
from azureml.pipeline.steps import PythonScriptStep, EstimatorStep
from azureml.pipeline.core.run import PipelineRun
from azureml.core.authentication import InteractiveLoginAuthentication
from azureml.interpret import ExplanationClient
import azureml.core
print("azureml.core version:", azureml.core.__version__)

### Workspace

# load the workspace
ws = Workspace.from_config()

### Experiment

# choose an experiment name
experiment = Experiment(ws, 'automl-binary-classification-SDSHackathon')

# dataset = pd.read_parquet("../../../data/platinum/dataset.parquet")
# dataset = pd.read_parquet("../../../data/silver/ultraPlusSpotify.parquet")
dataset = pd.read_parquet("../../../data/silver/ultraPlusSpotifyplusPDF.parquet")

# dataset.columns

# dataset.columns

dataset.columns

# dataset.sample(3)

# dataset.iloc[[50,41,13]]

dataset.iloc[[50,41,13]]

# dataset.drop(["#","Artiste_x","Titre","fileDate","freq","album","Artiste_y","release_date","time_signature"], axis=1, inplace=True)
# dataset.drop(["#","fileDate","Artiste","Titre","freq","CleanedTitle","id"], axis=1, inplace=True)
dataset.drop(["index","#_x","#_y","fileDate_x","fileDate_y","Artiste","Titre","freq","CleanedTitle","id","Artiest","Titel","Label","Distributeur","HP","W"], axis=1, inplace=True)

# we propose to use these features:
dataset.iloc[[50,41,13]]

dataset["VW"].unique()

dataset['VW'] = dataset['VW'].replace(np.nan, '50')

dataset['VW'] = dataset['VW'].replace('NEW', '50')

dataset = dataset.astype({'VW':'int'})

# dataset["True_Target"].sum()

dataset["True_Target"].sum()

dataset.shape

# we are not going to split for now... use the whole dataset

## Train

automl_settings = {
    "enable_early_stopping":True,
    "experiment_timeout_hours":0.25,
    "iterations":25, # number of runs
    "iteration_timeout_minutes":5,
    "max_concurrent_iterations":1,
    "max_cores_per_iteration":-1,
    #"experiment_exit_score":0.9920,
    "model_explainability":True,
    "n_cross_validations":5,
    "primary_metric":'precision_score_weighted', # we could use recall_score_weighted, AUC_weighted
    "featurization":'auto',
    "verbosity":logging.INFO, # {INFO, DEBUG, CRITICAL, ERROR, WARNING} -- debug_log=<*.log>
}

automl_config = AutoMLConfig(task='classification',
                             debug_log='automl_errors.log',
                             compute_target='local', # {training_cluster or 'local'}
                             #blacklist_models=['KNN','LinearSVM'],
                             enable_onnx_compatible_models=False,
                             training_data=dataset,
                             label_column_name="True_Target", # the name of the target variable
                             **automl_settings
                            )
# ouputs "model.pkl" and "automl_errors.log"

automl_run = experiment.submit(automl_config, show_output=True)

### explore the best pipeline

RunDetails(automl_run).show()
automl_run.wait_for_completion() # get more parameter info

### select best pipeline 

best_run, fitted_model = automl_run.get_output()

### inspect model properties

# pipeline steps
for step in fitted_model.named_steps:
    print(step)

# model properties
fitted_model.named_steps

# show all metrics
best_run.get_metrics()

### quick feature importance checks

client = ExplanationClient.from_run(best_run)
engineered_explanations = client.download_model_explanation(raw=False)
feature_importance = engineered_explanations.get_feature_importance_dict() # get model feature importance values
columns = ["modelFeatureImportance_name", "modelFeatureImportance_value"]
pd.DataFrame(list(feature_importance.items()), columns=columns)

## Register

####  prepare the scoring script, environment file and model

# get the score and environment files
model_name = best_run.properties['model_name'] # score.py script will look for the name of the registered model

# make a local copy of the best scoring script, environment file and the model file
script_file_name = 'inference/score.py'
conda_env_file_name = 'inference/env.yml'
model_pickle_file_name = 'inference/model.pkl'
# model_onnx_file_name = 'inference/model.onnx'
best_run.download_file('outputs/scoring_file_v_1_0_0.py', script_file_name)
best_run.download_file('outputs/conda_env_v_1_0_0.yml', conda_env_file_name)
best_run.download_file('outputs/model.pkl', model_pickle_file_name)
# best_run.download_file('outputs/model.onnx', model_onnx_file_name)

! cat inference/env.yml

! cat inference/score.py

### Register the model 

#### we use the local /path/model folder with Model.register()

model = Model.register(workspace=ws,
                       model_name=model_name, # registered model name used in scoring script init()
                       model_framework=Model.Framework.SCIKITLEARN, # {TensorFlow, ScikitLearn, Onnx, Custom}
                       model_framework_version='0.22.2',
                       model_path='inference/model.pkl', # local file {'model.pkl', 'model.onnx'}
                       tags={'Training context': 'autoML Training'},
                       properties={'AUC': best_run.get_metrics()['AUC_weighted'],
                                   'Accuracy': best_run.get_metrics()['accuracy']},
                       description="Classification model to predict diabetes")

### Deploy model as webservice (ACI)

#### this deplayment can take 15-20min... (using smallest compute 1CPU, 1GB mem)

%%time
# Configure the scoring environment
service_name = "automl-projname-service" # only lowercase letters, numbers, or dashes

# Remove any existing service under the same name
try:
    Webservice(ws, service_name).delete()
except WebserviceException:
    print('"' + service_name + '" does not exist, creating the webservice...')

myenv = Environment.from_conda_specification(name="myenv", file_path=conda_env_file_name)
inference_config = InferenceConfig(entry_script=script_file_name, environment=myenv)

deployment_config = AciWebservice.deploy_configuration(cpu_cores=1,
                                                       memory_gb=1)

# build container from environment, start webservice ACI and deploy inference scrips 
service = Model.deploy(ws, service_name, [model], inference_config, deployment_config)
service.wait_for_deployment(show_output=True)

# get webservice logs
print(service.get_logs())

Send a HTTP triggered webrequest with testdata to the model for a prediction value.  
In this example we test a person is inTop50 (1) or not-Top50 (0).  
The data must be a list of 11 features to predict a binary classification.  

# input_sample = pd.DataFrame({"VW": pd.Series([0], dtype="int64"), "acousticness": pd.Series([0.0], dtype="float64"), "danceability": pd.Series([0.0], dtype="float64"), "duration_ms": pd.Series([0], dtype="int64"), "energy": pd.Series([0.0], dtype="float64"), "instrumentalness": pd.Series([0.0], dtype="float64"), "key": pd.Series([0], dtype="int64"), "liveness": pd.Series([0.0], dtype="float64"), "loudness": pd.Series([0.0], dtype="float64"), "mode": pd.Series([0], dtype="int64"), "popularity": pd.Series([0], dtype="int64"), "speechiness": pd.Series([0.0], dtype="float64"), "tempo": pd.Series([0.0], dtype="float64"), "valence": pd.Series([0.0], dtype="float64")})
# output_sample = np.array([0])

# get webservice URI
endpoint = service.scoring_uri

# raw test data
 
# rawdata = [[50, 2.55e-05, 0.502, 301920, 0.912, 0.000173, 1, 0.106, -4.556, 1, 80, 0.0564, 116.761, 0.72],
#            [50, 6, 0, 148, 58, 11, 179, 39, 0.16, 45]]
rawdata = [[50, 2.55e-05, 0.502, 301920, 0.912, 0.000173, 1, 0.106, -4.556, 1, 80, 0.0564, 116.761, 0.72]]

print("URI: " + endpoint)
print("Body: " + json.dumps({"data": rawdata})) # convert array to a serialized JSON formatted string object

response = requests.post(endpoint, json={"data": rawdata})
response.json()

# service.delete()
